+++ eir/target_setup/target_label_setup.py
@@ -434,7 +434,7 @@ def process_survival_output(
     df_time = pl.read_csv(
         tabular_info.file_path,
         columns=["ID", time_column],
-    ).with_columns([pl.col("ID").cast(pl.Utf8), pl.col(time_column).cast(pl.Float64)])
+    ).with_columns([pl.col("ID").cast(pl.Utf8), pl.col(time_column).cast(pl.Float32)])
 
     df_time_train = df_time.filter(pl.col("ID").is_in(train_ids))
     df_time_valid = df_time.filter(pl.col("ID").is_in(valid_ids))
